pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lob-simulator"
version = "0.1.0"
description = "Toy LOB simulator — market microstructure + game-theory execution model"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "matplotlib>=3.8",
    "tqdm>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.hatch.build.targets.wheel]
packages = ["src/lob_simulator"]
